adjust project for building using CMake
