--- Software/bsw/gen/CanIf_Cfg.h
@@ -1,38 +1,0 @@
-/*******************************************************************************
-**                                                                            **
-**  Copyright (C) AUTOSarZs olc (2019)		                                  **
-**                                                                            **
-**  All rights reserved.                                                      **
-**                                                                            **
-**  This document contains proprietary information belonging to AUTOSarZs     **
-**  olc . Passing on and copying of this document, and communication          **
-**  of its contents is not permitted without prior written authorization.     **
-**                                                                            **
-********************************************************************************
-**                                                                            **
-**  FILENAME     : CanIf_Cfg			                                      **
-**                                                                            **
-**  VERSION      : 1.0.0                                                      **
-**                                                                            **
-**  DATE         : 2019-10-20                                                 **
-**                                                                            **
-**  VARIANT      : Variant PB                                                 **
-**                                                                            **
-**  PLATFORM     : TIVA C		                                              **
-**                                                                            **
-**  AUTHOR       : AUTOSarZs-DevTeam	                                      **
-**                                                                            **
-**  VENDOR       : AUTOSarZs OLC	                                          **
-**                                                                            **
-**                                                                            **
-**  DESCRIPTION  : CAN Interface source file                                  **
-**                                                                            **
-**  SPECIFICATION(S) : Specification of CAN Inteface, AUTOSAR Release 4.3.1   **
-**                                                                            **
-**  MAY BE CHANGED BY USER : no                                               **
-**                                                                            **
-*******************************************************************************/
-#ifndef __CANIF_CFG_H__
-#define __CANIF_CFG_H__
-
-#endif /* __CANIF_CFG_H__ */
