--- Software/bsw/gen/Can_PBcfg.h
@@ -1,6 +1,0 @@
-#ifndef INCLUDES_CAN_PBCFG_H_
-#define INCLUDES_CAN_PBCFG_H_
-
-
-
-#endif
