+++ Software/bsw/static/mcal/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.10)
+
+project(autosarz)
+
+message( STATUS "Building MCAL Components")
+add_subdirectory("${can_swc}")
