+++ Software/bsw/static/mcal/can_swc/config/Can_PBcfg.h
@@ -1,0 +1,6 @@
+#ifndef INCLUDES_CAN_PBCFG_H_
+#define INCLUDES_CAN_PBCFG_H_
+
+
+
+#endif
