+++ Software/bsw/static/mcal/can_swc/core/README.txt
@@ -1,0 +1,34 @@
+/*******************************************************************************
+**                                                                            **
+**  Copyright (C) AUTOSarZs olc (2019)		                                  **
+**                                                                            **
+**  All rights reserved.                                                      **
+**                                                                            **
+**  This document contains proprietary information belonging to AUTOSarZs     **
+**  olc . Passing on and copying of this document, and communication          **
+**  of its contents is not permitted without prior written authorization.     **
+**                                                                            **
+********************************************************************************
+**                                                                            **
+**  FOLDERNAME     : CAN         			                                  **
+**                                                                            **
+**  VERSION      : 1.0.0                                                      **
+**                                                                            **
+**  DATE         : 2019-09-22                                                 **
+**                                                                            **
+**  VARIANT      : Variant PB                                                 **
+**                                                                            **
+**  PLATFORM     : TIVA C TC431x                                              **
+**                                                                            **
+**  AUTHOR       : AUTOSarZs-DevTeam	                                      **
+**                                                                            **
+**  VENDOR       : AUTOSarZs OLC	                                          **
+**                                                                            **
+**                                                                            **
+**  DESCRIPTION  : CAN Driver source file                                     **
+**                                                                            **
+**  SPECIFICATION(S) : Specification of CAN Driver, AUTOSAR Release 4.3.1     **
+**                                                                            **
+**  MAY BE CHANGED BY USER : no                                               **
+**                                                                            **
+*******************************************************************************/
