+++ Software/bsw/static/mcal/can_swc/tools/cmake/can_swc_compile_options.cmake
@@ -1,0 +1,6 @@
+set(
+    can_swc_compile_options
+
+
+
+)
