+++ Software/bsw/static/mcal/can_swc/tools/cmake/can_swc_defines.cmake
@@ -1,0 +1,5 @@
+set(
+    can_swc_defines
+
+
+)
