+++ Software/bsw/static/mcal/can_swc/tools/cmake/can_swc_includes.cmake
@@ -1,0 +1,11 @@
+set(
+
+    can_swc_includes
+
+    core
+    config
+
+    ${Infrastructure_inc}
+    ${platform_inc}
+    ${stubs}
+)
