+++ Software/bsw/static/mcal/can_swc/tools/cmake/can_swc_sources.cmake
@@ -1,0 +1,7 @@
+set(
+    can_swc_sources_list
+
+    core/Can.c
+    config/Can_PBcfg.c
+
+)
