--- Software/swc/app/README.txt
@@ -1,34 +1,0 @@
-/*******************************************************************************
-**                                                                            **
-**  Copyright (C) AUTOSarZs olc (2019)		                                  **
-**                                                                            **
-**  All rights reserved.                                                      **
-**                                                                            **
-**  This document contains proprietary information belonging to AUTOSarZs     **
-**  olc . Passing on and copying of this document, and communication          **
-**  of its contents is not permitted without prior written authorization.     **
-**                                                                            **
-********************************************************************************
-**                                                                            **
-**  FOLDERNAME     : app         			                                  **
-**                                                                            **
-**  VERSION      : 1.0.0                                                      **
-**                                                                            **
-**  DATE         : 2019-09-22                                                 **
-**                                                                            **
-**  VARIANT      : NA		                                                  **
-**                                                                            **
-**  PLATFORM     : TIVA C TC431x                                              **
-**                                                                            **
-**  AUTHOR       : AUTOSarZs-DevTeam	                                      **
-**                                                                            **
-**  VENDOR       : AUTOSarZs OLC	                                          **
-**                                                                            **
-**                                                                            **
-**  DESCRIPTION  : CAN Driver source file                                     **
-**                                                                            **
-**  SPECIFICATION(S) : AUTOSarZs Test Code 								      **
-**                                                                            **
-**  MAY BE CHANGED BY USER : no                                               **
-**                                                                            **
-*******************************************************************************/
