+++ Software/swc/app_swc/tools/cmake/app_swc_compile_options.cmake
@@ -1,0 +1,6 @@
+set(
+    app_swc_compile_options
+
+
+
+)
