+++ Software/swc/app_swc/tools/cmake/app_swc_defines.cmake
@@ -1,0 +1,5 @@
+set(
+    app_swc_defines
+
+
+)
