+++ Software/swc/app_swc/tools/cmake/app_swc_includes.cmake
@@ -1,0 +1,14 @@
+set(
+
+    app_swc_includes
+
+    core
+    config
+
+    ${can_swc}/core
+    ${can_swc}/config
+    
+    ${Infrastructure_inc}
+    ${platform_inc}
+    ${stubs}
+)
