+++ Software/swc/app_swc/tools/cmake/app_swc_libs.cmake
@@ -1,0 +1,9 @@
+
+set(
+    app_swc_libs
+
+    "can_swc"
+    "-T ${CMAKE_PROJECT_DIR}/tm4c123gh6pz.lds"
+)
+
+
