--- Tools/README.txt
@@ -1,6 +1,0 @@
-Tool Folder : 
-Add any tool that help the developer to use the repo.
-
-Hint: Takecare or the size 
-
-Test Seesion 5
