+++ makefile
@@ -1,8 +1,10 @@
 include Makefile.variable
 
-CMAKE_VARIABLES := \
-	-DCMAKE_TOOLCHAIN_FILE=$(CMAKE_PROJECT_DIR)/toolchain.cmake \
-	-DTOOLCHAIN=$(TOOLCHAIN)
+CMAKE_VARIABLES =
+CMAKE_VARIABLES += -DCMAKE_TOOLCHAIN_FILE=$(CMAKE_PROJECT_DIR)/toolchain.cmake
+CMAKE_VARIABLES += -DTOOLCHAIN=$(TOOLCHAIN)
+CMAKE_VARIABLES += -DCMAKE_PROJECT_DIR=$(CMAKE_PROJECT_DIR)
+CMAKE_VARIABLES += -DSW_DIR=$(SW_DIR)
 
 autosar:
 ifeq ($(rebuild),1)
@@ -11,5 +13,5 @@ endif
 	@mkdir -p $(BUILD_DIR)
 	@echo "BUILD_DIR=$(BUILD_DIR)"
 	cd $(BUILD_DIR)/ && \
-	cmake -G "Unix Makefiles" -S $(CMAKE_PROJECT_DIR) -B $(BUILD_DIR) $(CMAKE_VARIABLES) && \
-	make AutosarApp install -j$(MAXIMUM_CPU_CORES)
+	cmake -G "Unix Makefiles" -S $(CURDIR) -B $(BUILD_DIR) $(CMAKE_VARIABLES) ;\
+	make app_swc install -j$(MAXIMUM_CPU_CORES)
